+++ Products/PluggableAuthService/plugins/tests/test_RequestTypeSniffer.py
@@ -22,20 +22,10 @@ from Products.PluggableAuthService.tests.conformance import \
 class RequestTypeSniffer(unittest.TestCase, IRequestTypeSniffer_conformance):
 
     def _getTargetClass(self):
-
         from Products.PluggableAuthService.plugins.RequestTypeSniffer \
             import RequestTypeSniffer
 
         return RequestTypeSniffer
 
     def _makeOne(self, id='test', *args, **kw):
-
         return self._getTargetClass()(id, *args, **kw)
-
-
-if __name__ == '__main__':
-    unittest.main()
-
-
-def test_suite():
-    return unittest.TestSuite((unittest.makeSuite(RequestTypeSniffer),))