+++ Products/PluggableAuthService/tests/test_MoreCaching.py
@@ -12,8 +12,6 @@
 #
 ##############################################################################
 
-import unittest
-
 from AccessControl.Permissions import view as View
 
 from Products.PluggableAuthService.interfaces.plugins import IExtractionPlugin
@@ -309,12 +307,3 @@ class CachingTests(pastc.PASTestCase):
 
         groups.removePrincipalFromGroup(user_id, group_id)
         self.assertCacheStats(0, 0, 0)
-
-
-def test_suite():
-    return unittest.TestSuite((
-        unittest.makeSuite(CachingTests),))
-
-
-if __name__ == '__main__':
-    unittest.main(defaultTest='test_suite')